pyproject.toml
[project]
name = "web-search-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.0.5",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
